AUT-5359: Move the PasskeysRetrieveResponse into shared
- Since we will also be fetching passkeys from account-management-api, it makes sense for this record to live in shared
