+++ shared/src/main/java/uk/gov/di/authentication/shared/entity/passkeys/PasskeysRetrieveResponse.java
@@ -1,4 +1,4 @@
-package uk.gov.di.authentication.frontendapi.entity.passkeys;
+package uk.gov.di.authentication.shared.entity.passkeys;
 
 import com.google.gson.annotations.Expose;
 import com.google.gson.annotations.SerializedName;
